backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=1.5.0
pyarrow>=12.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
joblib>=1.3.0
# IBM Db2 client (enables DB_MODE=db2 in backend)
ibm_db>=3.2.0
# Optional: watsonx.ai SDK
# ibm-watsonx-ai>=1.0.0
# Tests
pytest>=7.0.0
httpx>=0.25.0
